Gamemodemanagerbase cam, Deathmatch mode manager, playeractor changes
Added snow day camera to GamemodeManagerBase class to set up any cameras in any of the game modes.  If it is null it just skips init of the cam.
Created DeathmatchModeManager for the deathmatch game mode.  It contains feature parity to the old one but in a cleaner way.  Has no win screen as it is not in the game yet.
Changed the playeractor script to work with new manager.